backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# OpenAI & LangChain
openai>=1.10.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.16

# Vector Store
chromadb>=0.4.0

# Document Processing
pypdf>=3.0.0
python-docx>=1.0.0
tiktoken>=0.5.0

# Utilities
python-dotenv>=1.0.0
